restructure: rename views/ to api/
